Update README.md
Some of these documents have the identification in a mysql database, and the scripts connect to it
and read the identification for this file. Putting in the correct directory.

It begins with a directory full of pdfs, and at the end, we would have a directory with the files ordered in
directories which name is they identification.